improve perf - remove redundant call to Get-PSRepository
The existing PS commands that accept a repo (eg Install-Module)
will already throw a message saying that the repo supplied cannot
be found.

Calling `Get-PSRepository` for *each* PS repository takes approx.
3 seconds.

So given a requirement.psd1 file with 2 repositories, 6s will be
spent validating the repo exists. This time is a waste PS already
is checking for this condition.